add translation to SubjectViewer and unit tests for SingleVideoViewerContainer
