+++ packages/lib-classifier/package.json
@@ -67,7 +67,6 @@
     "@storybook/builder-webpack5": "~6.5.0",
     "@storybook/manager-webpack5": "~6.5.0",
     "@storybook/react": "~6.5.0",
-    "@storybook/testing-react": "~1.3.0",
     "@testing-library/dom": "~8.17.1",
     "@testing-library/react": "~12.1.2",
     "@testing-library/react-hooks": "~8.0.0",
