+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SingleVideoViewer/index.js
@@ -1,2 +1,2 @@
-export { default } from './SingleVideoViewerContainerFunctional'
+export { default } from './SingleVideoViewerContainer'
 export { default as SingleVideoViewer } from './SingleVideoViewer'
