+++ packages/lib-classifier/src/translations/en/components.json
@@ -156,6 +156,7 @@
     "previous": "Go to previous step"
   },
   "SubjectViewer": {
+    "loading": "Loading a subject",
     "BarChartViewer": {
       "chartLabel": "Bar chart"
     },
@@ -190,6 +191,9 @@
         "previousFrameLabel": "Previous",
         "nextFrameLabel": "Next"
       }
+    },
+    "SingleVideoViewerContainer": {
+      "error": "There was an error loading the subject."
     },
     "VariableStarViewer": {
       "controls": "controls",
